Moved AbstractWeatherChange to trend package
