+++ src/main/java/com/mivek/model/PROBChange.java
@@ -1,6 +1,7 @@
 package com.mivek.model;
 
 import com.mivek.enums.WeatherChangeType;
+import com.mivek.model.trend.AbstractWeatherChange;
 
 /**
  * @author mivek
