+++ src/main/java/com/mivek/model/TEMPOChange.java
@@ -1,6 +1,7 @@
 package com.mivek.model;
 
 import com.mivek.enums.WeatherChangeType;
+import com.mivek.model.trend.AbstractWeatherChange;
 
 /**
  * Class representing a Tempo change in a taf message.
