+++ src/main/java/com/mivek/model/trend/AbstractWeatherChange.java
@@ -1,7 +1,8 @@
-package com.mivek.model;
+package com.mivek.model.trend;
 
 import com.mivek.enums.WeatherChangeType;
-import com.mivek.model.trend.AbstractTrend;
+import com.mivek.model.AbstractValidity;
+import com.mivek.model.TemperatureDated;
 
 /**
  * Class representing a weather change.
